+++ cedar-policy/CHANGELOG.md
@@ -83,6 +83,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   `ValidationError::location`. (#405)
 - `ValidationWarningKind` is now `non_exhaustive`, allowing future warnings to
   be added without a breaking change. (#404)
+- Improve schema parsing error messages when a cycle exists in the action
+  hierarchy to includes an action which is part of the cycle (#436, resolving #416).
 
 ### Fixed
 
@@ -96,6 +98,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   policy. (#412)
 - Fixed schema-based parsing of entity data that includes unknowns (for the
   `partial-eval` experimental feature). (#419)
+
+### Removed
+
+- Move `ValidationMode::Permissive` behind an experimental feature flag.
+  To continue using this feature you must enable the `permissive-validate` feature
+  flag.
 
 ## [2.4.2] - 2023-10-23
 Cedar Language Version: 2.1.2